Minor refactoring for tesing suite setup. Mainly added blueprints.py as well as setting up a testing configuration in config.py (and moving it into app). A lot of the blueprint infrastructure was already implemented however, so this refactor was not as big as initially thought. Also, slightly tightened the quiz generation prompt as I was running into some issues with it during testing.
