+++ app/__init__.py
@@ -1,21 +1,20 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
 from flask_login import LoginManager
-from config import Config
+from app.config import Config
 
 db = SQLAlchemy()
-migrate = Migrate()
 login_manager = LoginManager()
 login_manager.login_view = 'main.login'
 
-def create_app():
+def create_app(config):
     app = Flask(__name__)
-    app.config.from_object(Config)
-    db.init_app(app)
-    migrate.init_app(app, db)
-    login_manager.init_app(app)
-    from app.routes import main
+    app.config.from_object(config)
+
+    from app.blueprints import main
     app.register_blueprint(main)
-    from app import models
+
+    db.init_app(app)
+    login_manager.init_app(app)
+
     return app
