+++ app/blueprints.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+main = Blueprint('main', __name__)
+
+from app import routes, models
