--- config.py
@@ -1,12 +1,0 @@
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-class Config:
-    SECRET_KEY = os.environ.get('SECRET_KEY') or 'dev-secret-key'
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or 'sqlite:///notella.db'
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-     # Add when implementing AI quiz feature
-    OPENAI_API_KEY = os.environ.get('OPENAI_API_KEY')
